chore: move graph logic to different file
